refactor: move styled-components from components files to your own files and change folder structure
