+++ src/components/Header/index.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components/native'
+
+const Header = styled.Text`
+  font-family: Roboto;
+  margin: 8px 0 9px 0;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 16px;
+  text-align: center;
+  text-transform: uppercase;
+`
+
+export default Header
