+++ src/components/Loading/Loading.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Container, Indicator } from './Loading.styles'
+
+const Loading = () => (
+  <Container>
+    <Indicator color="#ff6c00" size="large" />
+  </Container>
+)
+
+export default Loading
