+++ src/components/Loading/Loading.styles.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components/native'
+
+export const Indicator = styled.ActivityIndicator``
+
+export const Container = styled.View`
+  position: absolute;
+  top: 0;
+  right: 0;
+  left: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.7);
+  justify-content: center;
+`
