+++ src/components/Modal/Modal.js
@@ -1,7 +1,15 @@
 import React from 'react'
 import Modal from 'react-native-modal'
-import styled from 'styled-components'
-import Button from './Button'
+import {
+  ModalInnerContainer,
+  ButtonContainer,
+  ChildrenContainer,
+  InnerComponentContainer,
+  ModalLeftButton,
+  ModalRightButton,
+  ModalText,
+  ModalTitle
+} from './Modal.styles'
 
 const CustomModal = props => (
   <Modal {...props} avoidKeyboard>
@@ -25,46 +33,5 @@ const CustomModal = props => (
     </ModalInnerContainer>
   </Modal>
 )
-
-const ChildrenContainer = styled.View``
-
-const InnerComponentContainer = styled.View``
-
-const ModalLeftButton = styled(Button)`
-  margin-right: 4px;
-  flex: 1;
-`
-
-const ModalRightButton = styled(Button)`
-  margin-left: 4px;
-  flex: 1;
-`
-
-const ModalInnerContainer = styled.View`
-  padding: 16px;
-  background-color: white;
-  border-radius: 16px;
-`
-
-const ModalTitle = styled.Text`
-  font-family: Roboto;
-  color: #ff6c00;
-  font-weight: bold;
-  font-size: 17px;
-`
-
-const ModalText = styled.Text`
-  font-family: Roboto;
-  margin-top: 12px;
-  color: #555;
-  font-size: 17px;
-`
-
-const ButtonContainer = styled.View`
-  margin-top: 12px;
-  flex-direction: row;
-  width: 100%;
-  justify-content: space-around;
-`
 
 export default CustomModal
