+++ src/components/ModalService/ModalService.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Modal from './Modal'
+import Modal from '../Modal'
 
 const ModalServiceContext = React.createContext(Promise.reject)
