+++ src/components/ModalService/index.js
@@ -1,0 +1,3 @@
+import { ModalServiceContext, ModalServiceProvider } from './ModalService'
+
+export { ModalServiceContext, ModalServiceProvider }
