+++ src/components/NewsLetterForm/NewsLetterForm.js
@@ -1,6 +1,5 @@
 import React from 'react'
-import styled from 'styled-components'
-
+import { Input } from './NewsLetterForm.styles'
 const NewsLetterForm = () => {
   return (
     <>
@@ -9,13 +8,5 @@ const NewsLetterForm = () => {
     </>
   )
 }
-
-const Input = styled.TextInput`
-  margin-top: 8px;
-  padding: 4px;
-  height: 32px;
-  border-color: #ccc;
-  border-width: 1px;
-`
 
 export default NewsLetterForm
