+++ src/components/NewsLetterForm/NewsLetterForm.styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const Input = styled.TextInput`
+  margin-top: 8px;
+  padding: 4px;
+  height: 32px;
+  border-color: #ccc;
+  border-width: 1px;
+`
