+++ src/components/NewsLetterForm/index.js
@@ -1,0 +1,3 @@
+import NewsLetterForm from './NewsLetterForm'
+
+export default NewsLetterForm
