+++ src/components/ProductDetails/ProductDetails.js
@@ -1,15 +1,29 @@
 import React, { useEffect, useContext } from 'react'
 import { ScrollView } from 'react-native'
 import { useSelector, useDispatch } from 'react-redux'
-import styled from 'styled-components'
-import { fetchProductDetails } from '../stores/products/productsAction'
-import formatCurrency from '../utils/formatCurrency'
-import Button from './Button'
-import Accordion from './Accordion'
-import Loading from './Loading'
-import NewsLetterForm from './NewsLetterForm'
-import { ModalServiceContext } from './ModalService'
-import removeHtmlTags from '../utils/removeHtmlTags'
+import { fetchProductDetails } from '../../stores/products/productsAction'
+import formatCurrency from '../../utils/formatCurrency'
+import Button from '../Button'
+import Accordion from '../Accordion'
+import Loading from '../Loading'
+import NewsLetterForm from '../NewsLetterForm'
+import { ModalServiceContext } from '../ModalService'
+import removeHtmlTags from '../../utils/removeHtmlTags'
+import {
+  ButtonContainer,
+  Container,
+  LeftContainer,
+  ProductBrand,
+  ProductCode,
+  ProductCurrentCost,
+  ProductImage,
+  ProductImageContainer,
+  ProductInfoContainer,
+  ProductName,
+  ProductPaymentMethod,
+  ProductPreviousCost,
+  RightContainer
+} from './ProductDetails.styles'
 
 const ProductDetails = ({ productSku }) => {
   const products = useSelector(state => state.products)
@@ -104,85 +118,5 @@ const ProductDetails = ({ productSku }) => {
   }
   return <Loading />
 }
-
-const Container = styled.View`
-  margin: 0 10px 10px 10px;
-  padding: 8px 10px;
-  background-color: #fff;
-`
-
-const ProductName = styled.Text`
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 20px;
-  line-height: 24px;
-  color: #212121;
-`
-
-const ProductImageContainer = styled.View`
-  align-items: center;
-  margin: 12px 0;
-`
-
-const ProductImage = styled.Image`
-  align-items: center;
-  height: 163px;
-  width: 163px;
-`
-
-const ProductInfoContainer = styled.View`
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: flex-end;
-  margin-top: 9px;
-`
-
-const ProductCode = styled.Text`
-  font-family: Roboto;
-  font-size: 14px;
-  line-height: 16px;
-  text-align: right;
-  color: #c4c4c4;
-`
-
-const ProductBrand = styled.Text`
-  font-family: Roboto;
-  font-weight: 500;
-  font-size: 20px;
-  line-height: 24px;
-  text-align: right;
-  color: #212121;
-`
-
-const LeftContainer = styled.View``
-const RightContainer = styled.View``
-
-const ProductPreviousCost = styled.Text`
-  font-family: Roboto;
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 18px;
-  text-decoration-line: line-through;
-  color: #cccccc;
-`
-
-const ProductCurrentCost = styled.Text`
-  font-family: Roboto;
-  font-weight: 500;
-  font-size: 24px;
-  color: #ff6c00;
-`
-
-const ProductPaymentMethod = styled.Text`
-  font-family: Roboto;
-  font-weight: normal;
-  font-size: 17px;
-  color: #6f6f6f;
-`
-
-const ButtonContainer = styled.View`
-  margin: 13px -5px;
-`
 
 export default ProductDetails
