+++ src/components/ProductDetails/index.js
@@ -1,0 +1,3 @@
+import ProductDetails from './ProductDetails'
+
+export default ProductDetails
