+++ src/components/ProductList/ProductList.styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const Container = styled.ScrollView`
+  padding: 0 10px 10px 10px;
+`
+
+export const ListContainer = styled.View`
+  margin-bottom: 10px;
+`
