+++ src/components/ProductList/index.js
@@ -1,0 +1,3 @@
+import ProductList from './ProductList'
+
+export default ProductList
