+++ src/components/ProductListItem/ProductListItem.styles.js
@@ -1,0 +1,67 @@
+import styled from 'styled-components/native'
+
+export const Container = styled.View`
+  flex: 1;
+  flex-direction: row;
+  height: 145px;
+  padding: 5px;
+  margin-top: ${props => (props.isFirstItem ? '0' : '10px')};
+  background: #ffffff;
+  border-radius: 4px;
+  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.2);
+`
+
+export const LeftContent = styled.View`
+  width: 122px;
+  justify-content: space-between;
+`
+
+export const RightContent = styled.View`
+  flex: 2;
+  justify-content: space-between;
+`
+
+export const ProductCode = styled.Text`
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 12px;
+  line-height: 14px;
+  text-align: center;
+  color: #c4c4c4;
+`
+
+export const ProductName = styled.Text`
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 12px;
+  line-height: 14px;
+  color: #212121;
+`
+
+export const ProductCostContainer = styled.View``
+
+export const ProductPreviousCost = styled.Text`
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 11px;
+  line-height: 13px;
+  text-decoration-line: line-through;
+  color: #999999;
+`
+
+export const ProductCurrentCost = styled.Text`
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 16px;
+  color: #e66100;
+`
+
+export const ProductImage = styled.Image`
+  height: 108px;
+  width: 108px;
+`
