+++ src/components/ProductListItem/index.js
@@ -1,0 +1,3 @@
+import ProductListItem from './ProductListItem'
+
+export default ProductListItem
