+++ src/screens/ProductsListScreen.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import styled from 'styled-components'
 import ProductList from '../components/ProductList'
+import Header from '../components/Header'
 
 const ProductListScreen = ({ navigation }) => {
   return (
     <>
-      <ListHeader>Lista de Produtos</ListHeader>
+      <Header>Lista de Produtos</Header>
       <ProductList
         itemsPerRequest={3}
         onButtonPress={productSku => navigation.navigate('ProductDetails', { productSku })}
@@ -13,16 +13,5 @@ const ProductListScreen = ({ navigation }) => {
     </>
   )
 }
-
-const ListHeader = styled.Text`
-  font-family: Roboto;
-  margin: 8px 0 9px 0;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 14px;
-  line-height: 16px;
-  text-align: center;
-  text-transform: uppercase;
-`
 
 export default ProductListScreen
